pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools"]

[project]
name = "nodeseekmcp"
version = "0.0.1"
description = "NodeSeek MCP Server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "fastmcp>=2.11.3",
    "gunicorn>=23.0.0",
    "uvicorn-worker>=0.3.0",
    "httpx[socks]>=0.28.1",
    "feedparser>=6.0.11",
    "arrow>=1.3.0",
    "pendulum>=3.1.0",
    "sqlalchemy[asyncio]>=2.0.43",
    "supervisor>=4.2.5",
    "aiosqlite>=0.21.0",
    "apscheduler>=3.11.0",
    "jinja2>=3.1.6",
]

[project.scripts]
nodeseekmcp = 'nodeseekmcp.cli:cli'

[tool.setuptools.dynamic]
version = { attr = "nodeimage.__version__.__version__" }

[tool.setuptools.packages.find]
where = ["."]
include = ["nodeseekmcp*"]

[tool.setuptools.package-data]
"*" = ["*.*"]
